backend/pyproject.toml
[tool.poetry]
name = "property_rental_manager"
version = "0.1.0"
description = ""
authors = ["thunchakorn <s.thunchakorn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.8.0"}
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
pytz = "^2024.1"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
mypy = "^1.10.1"
pylint = "^3.2.5"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
